// playwright-e2e/enums/case-flags/case-level-flags.ts
enum CaseLevelFlags {
  COMPLEX = 'Complex Case',
  URGENT = 'Urgent case',
  POWER_OF_ARREST = 'Power of arrest with Police',
  WARRANT_OF_ARREST = 'Warrant of arrest conclusion/withdrawn',
  WELSH_FORMS_AND_COMS = 'Welsh forms and communications',
  SPECIAL_MEASURE = 'Special measure',
  ADJUSTMENT = 'Reasonable adjustment',
  VULNERABLE = 'Vulnerable user',
  CONFIDENTIAL = 'Confidential party/address',
  BEHAVIOUR = 'Unacceptable/disruptive customer behaviour',
  VEXATIOUS = 'Vexatious litigant',
  RESTRAINT = 'Civil restraint order',
  BANNING_ORDER = 'Banning order',
  EVIDENCE = 'Audio/Video Evidence',
  INTERPRETER = 'Language Interpreter',
  DEATH = 'Death of a Party',
  LITIGATION_FRIEND = 'Litigation friend',
  LACKING_CAPACITY = 'Lacking capacity',
  DETAINED = 'Detained individual',
  OTHER = 'Other',
}

export default CaseLevelFlags;